Refactor Button and Icon components with FontAwesome
The Button and Icon components were refactored to make use of FontAwesome SVG icons which now allow for more customizations and scalability across the app. With this change, icons are now being exported as IconDefinition, which provides a more detailed description of the icon's structure, improving its readability when used. Button now also takes an optional icon prop. This change also allows for buttons to have an inline icon next to the label, enhancing UI and UX.

The storybook for Button has also been updated to reflect these changes, adding an additional layer of visual testing for buttons with icons.

In addition, various FontAwesome packages were added to our dependencies. These provide a wide selection of icons to use, making the app more interactive and visually appealing.