Updated import style and added new block
Updated all string usage in "src/app/page.tsx" to double quotes and semicolon at the end of the import statement for consistency. In addition, a new file "src/app/blocks/page.tsx" was created and main block was added. The "src/app/globals.css" file was simplified by removing redundant CSS rules. Lastly, the README was updated with clearer instructions on how to start editing, and the text was restructured to have better line breaks for