Add Chromatic workflow, update `.gitignore` and make label properties dynamic
This commit adds a new workflow for Chromatic, updates the `.gitignore` file, and refactors the createMaterialTheme function in `main.ts`. The new Chromatic workflow allows push events to trigger Chromatic UI tests, helping us catch visual regressions earlier. The `.gitignore` file was updated to ignore the `build-storybook.log` file which is generated during Chromatic tests to reduce clutter. The createMaterialTheme function in `main.ts` was refactored to handle scenarios where not all color maps (primary, secondary, tertiary) are provided. This increases the flexibility of our theming capabilities. Also, the 'chromatic' script is removed from 'package.json' to place it in a separate GitHub Actions workflow for better visibility and maintainability.