fix: "Include package-lock.json in version control"
The .gitignore file was updated to stop ignoring package-lock.json and a large number of dependency changes was added to the package-lock.json file.

This was done to ensure that all developers working on this project have the exact same versions of all dependencies, thus reducing the chance of seeing bugs caused by versioning inconsistencies. Previously, package-lock.json was ignored, meaning that npm install could install differing versions of dependencies on different machines.