fix: Update icon import and repository URL
Replaced the import link for IconDefinition from '@fortawesome/pro-regular-svg-icons' to '@fortawesome/fontawesome-svg-core' in multiple files to properly bind to the SVG icons for better compatibility. Also updated the repository URL in release.config.js to point to the correct repository 'https://github.com/Udixio/UI.git'. This ensures the accurate repository is referenced for releases.