fix: Correct background color for links used as buttons and update Tailwind Material
Upgraded the Tailwind material from version 1.0.0-beta.4 to 1.0.0-beta.11 in package json and package-lock json. Notable changes in this version include an improvement in color handling for secondary and tertiary colors in the Material Theme class.

In addition, the button's class was adjusted to include 'inline-block'. This alteration regards the button's position in the flow layout, enhancing its display behavior to align with other inline items on the web page.